# Not to use for loop in convolution function