@okuta 2017-02-03
Thank you for your suggestion.

But, your measurement may be wrong.
CDUA launches kernel asynchronously.
Please use `cuda.cupy.cuda.get_elapsed_time`.