@beam2d 2017-02-03
AFAIK, there is no reason of using for loop here. The code is using it just because the code was simply ported from PyCUDA-based one at v1.3 and we didn't have tensordot for PyCUDA so that the old one didn't use it.