@fukatani 2017-02-03
Thanks, I didn't know the basis of cupy profiling.

I retried measurement.

using for loop: 0.14107 sec
tensordot: 0.00114 sec

I fixed measurement code as follows.

```python
import cupy
import time
from chainer.utils import conv

sy, sx = 1, 1
ph, pw = 0, 0

x = cupy.arange(300000).reshape(10, 3, 100, 100)
col = conv.im2col_gpu(x, 100, 100, sy, sx, ph, pw)
W = cupy.arange(90000).reshape(3, 3, 100, 100)

out_c, _, kh, kw = W.shape
n, c, h, w = x.shape

out_h = conv.get_conv_outsize(h, kh, sy, ph)
out_w = conv.get_conv_outsize(w, kw, sx, pw)

start = cupy.cuda.Event()
start.record()
W_mat = W.reshape(out_c, -1)
col_mats = col.reshape(n, -1, out_h * out_w)
y = cupy.empty((n, out_c, out_h, out_w), dtype=x.dtype)
y_mats = y.reshape(n, out_c, -1)
for i in range(n):
    y_mats[i] = W_mat.dot(col_mats[i])
end = cupy.cuda.Event()
end.record()
end.synchronize()

print(cupy.cuda.get_elapsed_time(start, end) / 1000)

start = cupy.cuda.Event()
start.record()
y = cupy.tensordot(col, W, ((1, 2, 3), (1, 2, 3))).astype(x.dtype, copy=False)
end = cupy.cuda.Event()
end.record()
end.synchronize()
print(cupy.cuda.get_elapsed_time(start, end) / 1000)

```