@fukatani 2017-02-03
Story changes, but I want profiling manual in Chainer and Cupy Reference Manual.
I'm sorry if I missed it.