@okuta 2017-02-08
Thank you new benchmark.

I wrote benchmark too.

result:
0.00244079162598
0.000277716808319

```
import cupy
import time
from chainer.utils import conv

sy, sx = 1, 1
ph, pw = 0, 0

x = cupy.arange(300000).reshape(10, 3, 100, 100)
col = conv.im2col_gpu(x, 100, 100, sy, sx, ph, pw)
W = cupy.arange(90000).reshape(3, 3, 100, 100)

out_c, _, kh, kw = W.shape
n, c, h, w = x.shape

out_h = conv.get_conv_outsize(h, kh, sy, ph)
out_w = conv.get_conv_outsize(w, kw, sx, pw)

def old():
    W_mat = W.reshape(out_c, -1)
    col_mats = col.reshape(n, -1, out_h * out_w)
    y = cupy.empty((n, out_c, out_h, out_w), dtype=x.dtype)
    y_mats = y.reshape(n, out_c, -1)
    for i in range(n):
        y_mats[i] = W_mat.dot(col_mats[i])
    return y

def new():
    return cupy.tensordot(col, W, ((1, 2, 3), (1, 2, 3))).astype(x.dtype, copy=False)

old()  # avoid initialize
new()  # avoid initialize

cnt = 100
stream = cupy.cuda.Stream.null
start = stream.record()
for i in range(cnt):
    old()
end = stream.record()
end.synchronize()

print(cupy.cuda.get_elapsed_time(start, end) / 1000 / cnt)

start = stream.record()
for i in range(cnt):
    new()
end = stream.record()
end.synchronize()
print(cupy.cuda.get_elapsed_time(start, end) / 1000 / cnt)
```