@mattya 2017-02-09
I also benchmarked the performance on ResNet50Layers.
The input shape is (16, 3, 256, 256).

| setup        | forward [sec]           | backward [sec]  | memory usage [MB] |
| ------------- |-------------:| -----:| -----:|
| your code      | 0.166      |  0.239 |  6800 |
| original code     | 0.198  | 0.265 | 6800 |
| (use cudnn) | 0.072 | 0.130 | 3900 |