@fukatani 2017-02-09
I learned a lot.
It is reasonable to get a difference about batch size (X10) and my benchmark is incorrect.

Resnet uses many conv layers, but the impact of this performance improvement is smaller than I expected.